requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.1
django-filter==22.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==1.5
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
virtualenv==20.14.1
whitenoise==6.3.0
